[REFACTOR] 컨트롤러 응답 매핑 로직 전담 매퍼로 분리
