+++ src/main/kotlin/com/coffee/api/cafe/presentation/adapter/in/restapi/CafeController.kt
@@ -5,6 +5,8 @@ import com.coffee.api.cafe.application.port.inbound.FindCafeArea
 import com.coffee.api.cafe.application.port.inbound.FindCafeDetails
 import com.coffee.api.cafe.application.port.inbound.FindRecommendCafe
 import com.coffee.api.cafe.presentation.adapter.`in`.restapi.dto.response.*
+import com.coffee.api.cafe.presentation.adapter.`in`.restapi.mapper.FindAllCafesResponseMapper
+import com.coffee.api.cafe.presentation.adapter.`in`.restapi.mapper.GetCafeDetailsResponseMapper
 import com.coffee.api.cafe.presentation.docs.CafeApi
 import com.coffee.api.common.support.response.ApiResponse
 import org.springframework.web.bind.annotation.GetMapping
@@ -29,29 +31,7 @@ class CafeController(
         @RequestParam(value = "area", required = false) area: String?,
     ): ApiResponse<FindAllCafesResponseWrapper> {
         val result = findCafe.invoke(FindCafe.Query(lastCafeId, area))
-
-        val cafes = result.cafes.map { cafe ->
-            FindAllCafesResponse(
-                cafeId = cafe.cafeId.value.toString(),
-                name = cafe.name,
-                nearestStation = cafe.nearestStation,
-                location = cafe.location,
-                price = cafe.price,
-                previewImages = cafe.previewImages,
-                tags = cafe.tags.map { tag ->
-                    TagResponse(
-                        id = tag.id.value.toString(),
-                        name = tag.name,
-                    )
-                },
-            )
-        }
-
-        val response = FindAllCafesResponseWrapper(
-            cafes = cafes,
-            hasNext = result.hasNext,
-        )
-
+        val response = FindAllCafesResponseMapper.toResponse(result)
         return ApiResponse.success(response)
     }
 
@@ -60,51 +40,7 @@ class CafeController(
         @PathVariable cafeId: UUID,
     ): ApiResponse<GetCafeDetailsResponse> {
         val result = findCafeDetails.invoke(FindCafeDetails.Query(cafeId))
-
-        val response = GetCafeDetailsResponse(
-            cafe = CafeResponse(
-                id = result.cafeDetails.cafe.id.value.toString(),
-                reasonForSelection = result.cafeDetails.cafe.reasonForSelection,
-                naverMapUrl = result.cafeDetails.cafe.naverMapUrl,
-                name = result.cafeDetails.cafe.name,
-                nearestStation = result.cafeDetails.cafe.nearestStation,
-                location = result.cafeDetails.cafe.location,
-                price = result.cafeDetails.cafe.price,
-                mainImageUrl = result.cafeDetails.cafe.mainImages,
-            ),
-            coffeeBean = CoffeeBeanResponse(
-                id = result.cafeDetails.coffeeBean.id.value.toString(),
-                description = result.cafeDetails.coffeeBean.description,
-                name = result.cafeDetails.coffeeBean.name,
-                engName = result.cafeDetails.coffeeBean.engName,
-                flavors = result.cafeDetails.coffeeBean.flavors.map { flavor ->
-                    FlavorResponse(
-                        flavor.displayName,
-                        flavor.category
-                    )
-                },
-                countryOfOrigin = result.cafeDetails.coffeeBean.countryOfOrigin,
-                roastingPoint = result.cafeDetails.coffeeBean.roastingPoint.toString(),
-            ),
-            menus = result.cafeDetails.menu.map { menu ->
-                MenuResponse(
-                    id = menu.id.value.toString(),
-                    name = menu.name,
-                    price = menu.price,
-                    imageUrl = menu.imageUrl,
-                    description = menu.description,
-                )
-            },
-            tags = result.cafeDetails.tag.map { tag ->
-                DetailTagResponse(
-                    id = tag.id.value.toString(),
-                    name = tag.name,
-                    imageUrl = tag.imageUrl
-                )
-            },
-            updatedAt = result.cafeDetails.updatedAt,
-        )
-
+        val response = GetCafeDetailsResponseMapper.toResponse(result)
         return ApiResponse.success(response)
     }
