+++ src/main/kotlin/com/coffee/api/cafe/presentation/adapter/in/restapi/dto/response/FindAllCafesResponse.kt
@@ -1,11 +1,6 @@
 package com.coffee.api.cafe.presentation.adapter.`in`.restapi.dto.response
 
 
-data class FindAllCafesResponseWrapper(
-    val cafes: List<FindAllCafesResponse>,
-    val hasNext: Boolean
-)
-
 data class FindAllCafesResponse(
     val cafeId: String,
     val name: String,
