+++ src/main/kotlin/com/coffee/api/cafe/presentation/adapter/in/restapi/dto/response/FindAllCafesResponseWrapper.kt
@@ -1,0 +1,6 @@
+package com.coffee.api.cafe.presentation.adapter.`in`.restapi.dto.response
+
+data class FindAllCafesResponseWrapper(
+    val cafes: List<FindAllCafesResponse>,
+    val hasNext: Boolean
+)
